+++ TOMB4/game/items.h
@@ -1,15 +1,15 @@
 #pragma once
 #include "../global/types.h"
 
-void InitialiseItemArray(size_t num);
-void KillItem(size_t item_num);
+void InitialiseItemArray(short num);
+void KillItem(short item_num);
 short CreateItem();
-void InitialiseItem(size_t item_num);
-void RemoveActiveItem(size_t item_num);
-void RemoveDrawnItem(size_t item_num);
-void AddActiveItem(size_t item_num);
-void ItemNewRoom(size_t item_num, short room_num);
-void InitialiseFXArray(size_t allocmem);
+void InitialiseItem(short item_num);
+void RemoveActiveItem(short item_num);
+void RemoveDrawnItem(short item_num);
+void AddActiveItem(short item_num);
+void ItemNewRoom(short item_num, short room_num);
+void InitialiseFXArray(long allocmem);
 short CreateEffect(short room_num);
 void KillEffect(short fx_num);
 void EffectNewRoom(short fx_num, short room_num);
